mpesa-service/requirements.txt
Django==5.2.2
mysqlclient

djangorestframework
rest_framework_simplejwt
corsheaders
requests
